requirements.txt
beautifulsoup4==4.11.1
discord.py==1.6.0
